% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ann_get_threads}
\alias{ann_get_threads}
\title{Number of threads used for index builds and queries}
\usage{
ann_get_threads()
}
\value{
Integer.
}
\description{
Number of threads used for index builds and queries
}
