% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{AnnoyIndex}
\alias{AnnoyIndex}
\title{Annoy index}
\description{
Random projection forest, as in Spotify's Annoy. More trees means better
recall, a larger index and a slower build. Raise \code{search_budget} for recall
at query time. No Manhattan.
}
\examples{
x <- generate_clustered_data(1000L, 16L)$data
idx <- AnnoyIndex$new(x, n_trees = 25L)
res <- idx$predict(x[1:5, ], k = 10L)

}
\references{
Bernhardsson, Annoy, \url{https://github.com/spotify/annoy}
}
\section{Super class}{
\code{\link[annsearchR:AnnIndex]{AnnIndex}} -> \code{AnnoyIndex}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{search_budget}}{Integer or \code{NULL}. Candidates inspected per query.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-AnnoyIndex-initialize}{\code{AnnoyIndex$new()}}
    \item \href{#method-AnnoyIndex-clone}{\code{AnnoyIndex$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="annsearchR" data-topic="AnnIndex" data-id="predict"><a href='../../annsearchR/html/AnnIndex.html#method-AnnIndex-predict'><code>AnnIndex$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="annsearchR" data-topic="AnnIndex" data-id="print"><a href='../../annsearchR/html/AnnIndex.html#method-AnnIndex-print'><code>AnnIndex$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="annsearchR" data-topic="AnnIndex" data-id="query_self"><a href='../../annsearchR/html/AnnIndex.html#method-AnnIndex-query_self'><code>AnnIndex$query_self()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="annsearchR" data-topic="AnnIndex" data-id="save"><a href='../../annsearchR/html/AnnIndex.html#method-AnnIndex-save'><code>AnnIndex$save()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnoyIndex-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-AnnoyIndex-initialize}{}}}
\subsection{\code{AnnoyIndex$new()}}{
  Build the index.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{AnnoyIndex$new(
  data,
  metric = c("euclidean", "sqeuclidean", "cosine"),
  n_trees = 25L,
  search_budget = NULL,
  seed = 42L,
  precision = c("float", "double")
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Numeric matrix or data.frame. Samples x features. An index
pointer is also accepted; \code{\link[=load_ann_index]{load_ann_index()}} uses that path.}
      \item{\code{metric}}{String. One of \code{c("euclidean", "sqeuclidean", "cosine")}.}
      \item{\code{n_trees}}{Integer. Number of trees in the forest.}
      \item{\code{search_budget}}{Integer or \code{NULL}. Candidates inspected per query.
\code{NULL} uses \code{k * n_trees * 20}. Can be changed after the build.}
      \item{\code{seed}}{Integer. Fixes the random hyperplanes.}
      \item{\code{precision}}{String. \code{"float"} stores the data as f32, \code{"double"}
as f64.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnoyIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnoyIndex-clone}{}}}
\subsection{\code{AnnoyIndex$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{AnnoyIndex$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
