% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_index.R
\name{.is_ptr}
\alias{.is_ptr}
\title{Is this an external pointer, alive or dead}
\usage{
.is_ptr(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean.
}
\description{
Is this an external pointer, alive or dead
}
\keyword{internal}
