% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_index.R
\name{predict.AnnIndex}
\alias{predict.AnnIndex}
\title{Query an index}
\usage{
\method{predict}{AnnIndex}(object, newdata, k = 15L, ...)
}
\arguments{
\item{object}{An \link{AnnIndex}.}

\item{newdata}{Numeric matrix or data.frame. Queries x features.}

\item{k}{Integer. Number of neighbours.}

\item{...}{Passed on to \verb{$predict()}: \code{return_dist}, \code{.verbose}.}
}
\value{
A list with \code{idx} and \code{dist}. See \link{AnnIndex}.
}
\description{
S3 wrapper around the \verb{$predict()} method of an \link{AnnIndex}.
}
