% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_balltree_build}
\alias{rs_balltree_build}
\title{Build a ball tree index}
\usage{
rs_balltree_build(data, metric, seed, precision)
}
\arguments{
\item{data}{Numeric matrix. Samples x features.}

\item{metric}{String. One of \code{c("euclidean", "cosine")}.}

\item{seed}{Integer. Random seed.}

\item{precision}{String. \code{"float"} or \code{"double"}.}
}
\value{
External pointer to the index.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \link{BallTreeIndex} instead.
}
\keyword{internal}
