% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_exhaustive_build}
\alias{rs_exhaustive_build}
\title{Build an exhaustive index}
\usage{
rs_exhaustive_build(data, metric, precision)
}
\arguments{
\item{data}{Numeric matrix. Samples x features.}

\item{metric}{String. One of \code{c("euclidean", "cosine", "manhattan")}.}

\item{precision}{String. \code{"float"} or \code{"double"}.}
}
\value{
External pointer to the index.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \link{ExhaustiveIndex} instead.
}
\keyword{internal}
