% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_get_threads}
\alias{rs_get_threads}
\title{Number of threads used for builds and queries}
\usage{
rs_get_threads()
}
\value{
Integer. The configured pool's size, or Rayon's global pool size.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \code{\link[=ann_get_threads]{ann_get_threads()}} instead.
}
\keyword{internal}
