% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_kdtree_build}
\alias{rs_kdtree_build}
\title{Build a kd forest index}
\usage{
rs_kdtree_build(data, metric, n_trees, seed, precision)
}
\arguments{
\item{data}{Numeric matrix. Samples x features.}

\item{metric}{String. One of \code{c("euclidean", "cosine", "manhattan")}.}

\item{n_trees}{Integer. Number of trees.}

\item{seed}{Integer. Random seed.}

\item{precision}{String. \code{"float"} or \code{"double"}.}
}
\value{
External pointer to the index.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \link{KdTreeIndex} instead.
}
\keyword{internal}
